Bug fix for LTSs without transitions
If a LTS has no transitions, the program originally would dereference a
transition past the last one, which produces an error message.  This has
been corrected so the last transition is dereferenced instead.

Also corrected a type error in artem_appl.h, which led to warnings about
type mismatch.

+++ libraries/atermpp/include/mcrl2/atermpp/detail/aterm_appl.h
@@ -110,7 +110,7 @@ class _aterm_appl_allocator
 {
 private:
   /// \returns The size (in bytes) of a class T with arity number of arguments placed at the end.
-  constexpr static std::uint64_t term_appl_size(std::uint64_t arity)
+  constexpr static std::size_t term_appl_size(std::size_t arity)
   {
     return sizeof(T) + (arity - 1) * sizeof(aterm);
   }
